+++ src/xkb/xkbcomp/types.rs
@@ -27,7 +27,7 @@ pub mod stdint_intn_h {
 pub mod stdint_uintn_h {
     pub type uint8_t = __uint8_t;
     pub type uint16_t = __uint16_t;
-    pub type uint32_t = __uint32_t;
+    pub type u32 = __uint32_t;
     use super::types_h::{__uint16_t, __uint32_t, __uint8_t};
 }
 pub mod __stddef_size_t_h {
@@ -173,12 +173,12 @@ pub mod xkbcommon_h {
     pub const XKB_LOG_LEVEL_WARNING: xkb_log_level = 30;
     pub const XKB_LOG_LEVEL_ERROR: xkb_log_level = 20;
     pub const XKB_LOG_LEVEL_CRITICAL: xkb_log_level = 10;
-    pub type xkb_layout_index_t = uint32_t;
-    pub type xkb_keycode_t = uint32_t;
-    pub type xkb_mod_mask_t = uint32_t;
-    pub type xkb_mod_index_t = uint32_t;
-    pub type xkb_keysym_t = uint32_t;
-    pub type xkb_level_index_t = uint32_t;
+    pub type xkb_layout_index_t = u32;
+    pub type xkb_keycode_t = u32;
+    pub type xkb_mod_mask_t = u32;
+    pub type xkb_mod_index_t = u32;
+    pub type xkb_keysym_t = u32;
+    pub type xkb_level_index_t = u32;
     pub type xkb_layout_out_of_range_policy = ::core::ffi::c_uint;
     pub const XKB_LAYOUT_OUT_OF_RANGE_REDIRECT: xkb_layout_out_of_range_policy = 2;
     pub const XKB_LAYOUT_OUT_OF_RANGE_CLAMP: xkb_layout_out_of_range_policy = 1;
@@ -194,8 +194,8 @@ pub mod xkbcommon_h {
     pub const XKB_STATE_MODS_LOCKED: xkb_state_component = 4;
     pub const XKB_STATE_MODS_LATCHED: xkb_state_component = 2;
     pub const XKB_STATE_MODS_DEPRESSED: xkb_state_component = 1;
-    pub type xkb_layout_mask_t = uint32_t;
-    pub type xkb_led_index_t = uint32_t;
+    pub type xkb_layout_mask_t = u32;
+    pub type xkb_led_index_t = u32;
     pub type xkb_keymap_format = ::core::ffi::c_uint;
     pub const XKB_KEYMAP_FORMAT_TEXT_V2: xkb_keymap_format = 2;
     pub const XKB_KEYMAP_FORMAT_TEXT_V1: xkb_keymap_format = 1;
@@ -203,7 +203,7 @@ pub mod xkbcommon_h {
     pub const XKB_KEYMAP_COMPILE_STRICT_MODE: xkb_keymap_compile_flags = 1;
     pub const XKB_KEYMAP_COMPILE_NO_FLAGS: xkb_keymap_compile_flags = 0;
     use super::context_h::xkb_context;
-    use super::stdint_uintn_h::uint32_t;
+    use super::stdint_uintn_h::u32;
     extern "C" {
         pub fn xkb_context_get_log_verbosity(context: *mut xkb_context) -> ::core::ffi::c_int;
     }
@@ -978,11 +978,11 @@ pub mod text_h {
     #[repr(C)]
     pub struct LookupEntry {
         pub name: *const ::core::ffi::c_char,
-        pub value: uint32_t,
+        pub value: u32,
     }
     use super::context_h::xkb_context;
     use super::keymap_h::{mod_type, xkb_mod_set};
-    use super::stdint_uintn_h::uint32_t;
+    use super::stdint_uintn_h::u32;
     use super::xkbcommon_h::xkb_mod_mask_t;
     extern "C" {
         pub fn ModMaskText(
@@ -1019,7 +1019,7 @@ pub mod xkbcomp_priv_h {
     pub struct pending_computation {
         pub expr: *mut ExprDef,
         pub computed: bool,
-        pub value: uint32_t,
+        pub value: u32,
     }
     #[derive(Copy, Clone)]
     #[repr(C)]
@@ -1119,7 +1119,7 @@ pub mod xkbcomp_priv_h {
         xkb_message_code, XKB_ERROR_EXPECTED_ARRAY_ENTRY, XKB_LOG_VERBOSITY_MINIMAL,
     };
     use super::stdbool_h::false_0;
-    use super::stdint_uintn_h::{uint32_t, uint8_t};
+    use super::stdint_uintn_h::{u32, uint8_t};
     use super::text_h::LookupEntry;
     use super::xkbcommon_h::{xkb_layout_index_t, XKB_LOG_LEVEL_ERROR};
     extern "C" {
@@ -1384,7 +1384,7 @@ pub use self::messages_codes_h::{
 };
 pub use self::stdbool_h::{false_0, true_0};
 pub use self::stdint_intn_h::{int16_t, int32_t, int64_t, int8_t};
-pub use self::stdint_uintn_h::{uint16_t, uint32_t, uint8_t};
+pub use self::stdint_uintn_h::{u32, uint16_t, uint8_t};
 use self::stdlib_h::{calloc, free, realloc};
 use self::string_h::memset;
 pub use self::text_h::{LookupEntry, ModMaskText};
